Side quest to remove OAuth2NetworkManager and LabsLoginController
