+++ PennMobile/General/Protocols.swift
@@ -11,6 +11,49 @@ import CoreLocation
 import LocalAuthentication
 import PennMobileShared
 
+import CommonCrypto
+import CryptoKit
+
+enum SHA256Encoding {
+    case base64
+    case hex
+}
+
+protocol SHA256Hashable {}
+
+extension SHA256Hashable {
+    func hash(string: String, encoding: SHA256Encoding) -> String {
+        let inputData = Data(string.utf8)
+#if canImport(CryptoKit)
+        let digest = SHA256.hash(data: inputData)
+        switch encoding {
+        case .base64:
+            return Data([UInt8](digest.makeIterator())).base64EncodedString()
+        case .hex:
+            return digest.compactMap { String(format: "%02x", $0) }.joined()
+        }
+#else
+        return commonCryptoHash(inputData: inputData, encoding: encoding)
+#endif
+    }
+
+    private func commonCryptoHash(inputData: Data, encoding: SHA256Encoding) -> String {
+        // https://www.agnosticdev.com/content/how-use-commoncrypto-apis-swift-5
+        var digest = [UInt8](repeating: 0, count: Int(CC_SHA256_DIGEST_LENGTH))
+        _ = inputData.withUnsafeBytes {
+            CC_SHA256($0.baseAddress, UInt32(inputData.count), &digest)
+        }
+
+        switch encoding {
+        case .base64:
+            return Data([UInt8](digest.makeIterator())).base64EncodedString()
+        case .hex:
+            return digest.compactMap { String(format: "%02x", $0) }.joined()
+        }
+    }
+}
+
+
 protocol IndicatorEnabled {}
 
 extension IndicatorEnabled where Self: UIViewController {
