+++ PennMobile/Setup + Navigation/ControllerModel.swift
@@ -55,7 +55,6 @@ class ControllerModel: NSObject {
         vcDictionary[.dining] = DiningViewControllerSwiftUI()
         vcDictionary[.studyRoomBooking] = GSRTabController()
         vcDictionary[.laundry] = LaundryTableViewController()
-        vcDictionary[.more] = MoreViewController()
         vcDictionary[.map] = MapViewController()
         vcDictionary[.news] = NewsViewController()
         vcDictionary[.about] = AboutViewController()
@@ -63,7 +62,6 @@ class ControllerModel: NSObject {
         vcDictionary[.preferences] = PreferencesViewController()
         vcDictionary[.privacy] = PrivacyViewController()
         vcDictionary[.courseSchedule] = CoursesViewController()
-        vcDictionary[.pacCode] = PacCodeViewController()
         vcDictionary[.courseAlerts] = CourseAlertController()
         vcDictionary[.headlineNews] = NativeNewsViewController()
     }
@@ -78,20 +76,6 @@ class ControllerModel: NSObject {
     var tabFeatures: [Feature] {
         get {
             return UserDefaults.standard.getTabPreferences()
-        }
-    }
-
-    // Features in MoreViewController:
-    var moreFeatures: [Feature] {
-        get {
-            let tabPreferences = UserDefaults.standard.getTabPreferences()
-            return dynamicFeatures.filter { !tabPreferences.contains($0) } + [.about]
-        }
-    }
-
-    var moreIcons: [UIImage] {
-        get {
-            return moreFeatures.map { featureIcons[$0] ?? #imageLiteral(resourceName: "logo-small") }
         }
     }
