+++ PennMobile/Setup + Navigation/Features.swift
@@ -125,7 +125,6 @@ let features: [AppFeature] = [
         MarketplaceView()
             .navigationTitle(Text("Marketplace"))
     },
-    AppFeature(.pac, shortName: "PAC", longName: "PAC Code", color: .grey5, image: .system("lock"), controller: PacCodeViewController.self),
     AppFeature(.ticketScanner, shortName: "Scanner", longName: "Ticket Scanner", color: .basePurple, image: .system("qrcode.viewfinder")) {
         ScannerView()
             .navigationTitle(Text("Scanner"))
