+++ PennMobile/Setup + Navigation/TabBarController.swift
@@ -111,11 +111,3 @@ extension NewsViewController: TabBarShowable {
         return UITabBarItem(title: "News", image: normalImage, selectedImage: selectedImage)
     }
 }
-
-extension MoreViewController: TabBarShowable {
-    func getTabBarItem() -> UITabBarItem {
-        let normalImage = UIImage(named: "More_Grey")
-        let selectedImage = UIImage(named: "More_Blue")
-        return UITabBarItem(title: "More", image: normalImage, selectedImage: selectedImage)
-    }
-}
